+++ src/components/calc/productSummary.tsx
@@ -4,6 +4,7 @@ import {getProductPrice} from "@Utils/calculate";
 import {TiersComponent} from "./tiers";
 import {Product} from "@Utils/models/models";
 import {Price} from "../price/price";
+import {InputField} from "../form/fields/input";
 
 type Props = {
   product: Product,
@@ -17,6 +18,16 @@ export const ProductSummary = function(props: Props) {
     props.onChange && props.onChange(value);
   }
 
+  let error = '';
+
+  if(props.count <= 0){
+    error = 'Количество не может быть менее одного';
+  }
+
+  if(!props.count){
+    error = 'Количество не может быть пустым';
+  }
+
   return (
     <Form.Group as={Row} controlId="product" key={props.product.name}>
       <Form.Label column="sm" sm="2">{props.product.label}</Form.Label>
@@ -24,24 +35,23 @@ export const ProductSummary = function(props: Props) {
         <Price price={getProductPrice(props.product, props.count)} />
       </Col>
       <Col sm="2">
-        <InputGroup size="sm">
-          <Form.Control
-            type="number"
-            onChange={(e) => handleChange(e.target.value)}
-            min="0"
-            step="1"
-            value={props.count}
-            defaultValue={props.product.defaultCount}
-          />
-          <InputGroup.Prepend>
-            <InputGroup.Text>Шт</InputGroup.Text>
-          </InputGroup.Prepend>
-        </InputGroup>
+        <InputField
+          type="number"
+          name="count"
+          onChange={handleChange}
+          inputGroupLabel="Шт"
+          isEnabledInputGroup
+          smField={12}
+          min="0"
+          step="1"
+          value={props.count}
+          defaultValue={props.product.defaultCount}
+          error={error}
+         />
       </Col>
       <Col sm="3">
         <TiersComponent tiers={props.product.tiers} count={props.count} />
       </Col>
-
     </Form.Group>
   );
 }