+++ src/containers/calculator/calculator.tsx
@@ -22,6 +22,10 @@ export function Calculator(){
     return products.rows[0]?._id;
   });
 
+  useEffect(() => {
+    setValue(products.rows[0]?._id)
+  }, [products]);
+
   const options = products.rows.map((obj: Product) => ({label: obj.label, value: obj._id}));
   let product = products.rows.find((obj: Product) => obj._id === value);